Put all code (including Stuart-Landau) in kuramoto.c
