+++ C/demo/Makefile
@@ -16,7 +16,6 @@ INCDIR ?= $(PREFIX)/include/kuramoto
 BINDIR ?= $(PREFIX)/bin
 
 CC ?= gcc
-INSTALL = install -D
 
 # If you don't have Gnuplot installed, omit the  -D_HAVE_GNUPLOT flag
 CFLAGS = -std=c99 -march=native -O2 -Wall -Wextra -Winline -pedantic-errors -D_DEFAULT_SOURCE -D_HAVE_GNUPLOT
@@ -29,7 +28,7 @@ $(BIN): $(SRC) $(INC)
 	$(CC) $(SRC) $(CFLAGS) -I$(INCDIR) -Wl,-rpath=$(LIBDIR) -lkuramoto -lm -o $(BIN)
 
 install:
-	$(INSTALL) -m755 $(BIN) $(BINDIR)/$(BIN)
+	install -D -m755 $(BIN) $(BINDIR)/$(BIN)
 
 uninstall:
 	$(RM) $(BINDIR)/$(BIN)
